2014.06.12
bug fixed

+++ Requests/GRListingRequest.m
@@ -98,7 +98,7 @@
             NSUInteger  offset = 0;
             CFIndex     parsedBytes;
             uint8_t *bytes = (uint8_t *)[self.receivedData bytes];
-            int totalbytes = [self.receivedData length];
+            NSInteger totalbytes = [self.receivedData length];
             
             do {
                 CFDictionaryRef listingEntity = NULL;
